requirement.txt
aiofiles==22.1.0
alembic==1.4.3
amqp==5.0.1
APScheduler==3.6.3
arrow==0.17.0
async-timeout==3.0.1
atomicwrites==1.4.1
attrs==20.3.0
bcrypt==3.1.7
billiard==3.6.3.0
binaryornot==0.4.4
celery==5.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.6
contourpy==1.0.7
cookiecutter==1.7.2
cryptography==39.0.1
cycler==0.11.0
dnspython==1.16.0
ecdsa==0.14.1
email-validator==1.1.1
fastapi==0.63.0
fonttools==4.39.3
greenlet==2.0.1
gunicorn==20.0.4
h11==0.9.0
hiredis==1.1.0
httptools==0.5.0
idna==2.9
importlib-resources==5.12.0
iniconfig==1.1.1
install==1.3.4
Jinja2==2.11.2
jinja2-time==0.2.0
kiwisolver==1.4.4
kombu==5.0.2
loguru==0.5.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.7.1
numpy==1.24.2
packaging==20.8
passlib==1.7.4
peewee==3.15.4
Pillow==9.5.0
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==3.0.8
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.5.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.4.6
starlette==0.13.6
text-unidecode==1.3
toml==0.10.2
tzlocal==2.1
urllib3==1.25.10
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
win32-setctime==1.1.0
wincertstore==0.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
xxx==0.0.1
zipp==3.15.0
